Changing some import paths around autopkg
