+++ Code/autopkgcmd/searchcmd.py
@@ -16,7 +16,7 @@ from typing import List
 from urllib.parse import quote
 
 from autopkgcmd.opts import common_parse, gen_common_parser
-from autopkglib import log_err
+from autopkglib.common import log_err
 from autopkglib.github import (
     DEFAULT_SEARCH_USER,
     GitHubSession,
