+++ Code/autopkglib/__init__.py
@@ -21,59 +21,30 @@ import json
 import os
 import plistlib
 import pprint
-import re
 import subprocess
 import sys
 import traceback
-from copy import deepcopy
 from distutils.version import LooseVersion
-from typing import IO, Any, Dict, List, Optional, Union
+from typing import IO, Dict, List, Optional
 
-import appdirs
 import pkg_resources
 import yaml
-
-# Type for methods that accept either a filesystem path or a file-like object.
-FileOrPath = Union[IO, str, bytes, int]
-
-# Type for ubiquitous dictionary type used throughout autopkg.
-# Most commonly for `input_variables` and friends. It also applies to virtually all
-# usages of plistlib results as well.
-VarDict = Dict[str, Any]
-
-
-def is_mac():
-    """Return True if current OS is macOS."""
-    return "darwin" in sys.platform.lower()
-
-
-def is_windows():
-    """Return True if current OS is Windows."""
-    return "win32" in sys.platform.lower()
-
-
-def is_linux():
-    """Return True if current OS is Linux."""
-    return "linux" in sys.platform.lower()
-
-
-def log(msg, error=False):
-    """Message logger, prints to stdout/stderr."""
-    if error:
-        print(msg, file=sys.stderr)
-    else:
-        print(msg)
-
-
-def log_err(msg):
-    """Message logger for errors."""
-    log(msg, error=True)
-
-
-def autopkg_user_folder() -> str:
-    """Return a path string for the AutoPkg user folder"""
-    return os.path.abspath(os.path.expanduser("~/Library/AutoPkg"))
-
+from autopkglib.common import (
+    DEFAULT_SEARCH_DIRS,
+    DEFAULT_USER_OVERRIDES_DIR,
+    DEFAULT_RECIPE_MAP,
+    RE_KEYREF,
+    RECIPE_EXTS,
+    FileOrPath,
+    VarDict,
+    autopkg_user_folder,
+    is_linux,
+    is_mac,
+    is_windows,
+    log,
+    log_err,
+)
+from autopkglib.prefs import Preferences
 
 try:
     from CoreFoundation import (  # type: ignore
@@ -119,201 +90,6 @@ except ImportError:
     kCFPreferencesAnyUser = None
     kCFPreferencesCurrentUser = None
     kCFPreferencesCurrentHost = None
-
-APP_NAME = "Autopkg"
-BUNDLE_ID = "com.github.autopkg"
-DEFAULT_SEARCH_DIRS = [".", "~/Library/AutoPkg/Recipes", "/Library/AutoPkg/Recipes"]
-
-RE_KEYREF = re.compile(r"%(?P<key>[a-zA-Z_][a-zA-Z_0-9]*)%")
-
-# Supported recipe extensions
-RECIPE_EXTS = (".recipe", ".recipe.plist", ".recipe.yaml")
-
-
-class PreferenceError(Exception):
-    """Preference exception"""
-
-    pass
-
-
-class Preferences:
-    """An abstraction to hold all preferences."""
-
-    def __init__(self) -> None:
-        """Init."""
-        self.prefs: VarDict = {
-            "RECIPE_SEARCH_DIRS": DEFAULT_SEARCH_DIRS,
-            "RECIPE_OVERRIDE_DIRS": ["~/Library/AutoPkg/RecipeOverrides"],
-            "GITHUB_TOKEN_PATH": "~/Library/AutoPkg/gh_token",
-        }
-        # What type of preferences input are we using?
-        self.type: Optional[str] = None
-        # Path to the preferences file we were given
-        self.file_path: Optional[str] = None
-        # If we're on macOS, read in the preference domain first.
-        if is_mac():
-            self.prefs = self._get_macos_prefs()
-        else:
-            self.prefs = self._get_file_prefs()
-        if not self.prefs:
-            log_err(
-                "Preferences domain com.github.autopkg looks empty, using default preferences or file."
-            )
-
-    def _parse_json_or_plist_file(self, file_path) -> VarDict:
-        """Parse the file. Start with plist, then JSON."""
-        try:
-            with open(file_path, "rb") as f:
-                data = plistlib.load(f)
-            self.type = "plist"
-            self.file_path = file_path
-            return data
-        except Exception:
-            pass
-        try:
-            with open(file_path, "rb") as f:
-                data = json.load(f)
-                self.type = "json"
-                self.file_path = file_path
-                return data
-        except Exception:
-            pass
-        return {}
-
-    def __deepconvert_objc(self, object):
-        """Convert all contents of an ObjC object to Python primitives."""
-        value = object
-        if isinstance(object, NSNumber):
-            value = int(object)
-        elif isinstance(object, NSArray) or isinstance(object, list):
-            value = [self.__deepconvert_objc(x) for x in object]
-        elif isinstance(object, NSDictionary):
-            value = dict(object)
-            # RECIPE_REPOS is a dict of dicts
-            for k, v in value.items():
-                if isinstance(v, NSDictionary):
-                    value[k] = dict(v)
-        else:
-            return object
-        return value
-
-    def _get_macos_pref(self, key):
-        """Get a specific macOS preference key."""
-        value = self.__deepconvert_objc(CFPreferencesCopyAppValue(key, BUNDLE_ID))
-        return value
-
-    def _get_macos_prefs(self):
-        """Return a dict (or an empty dict) with the contents of all
-        preferences in the domain."""
-        prefs = {}
-
-        # get keys stored via 'defaults write [domain]'
-        user_keylist = CFPreferencesCopyKeyList(
-            BUNDLE_ID, kCFPreferencesCurrentUser, kCFPreferencesAnyHost
-        )
-
-        # get keys stored via 'defaults write /Library/Preferences/[domain]'
-        system_keylist = CFPreferencesCopyKeyList(
-            BUNDLE_ID, kCFPreferencesAnyUser, kCFPreferencesCurrentHost
-        )
-
-        # CFPreferencesCopyAppValue() in get_macos_pref() will handle returning the
-        # appropriate value using the search order, so merging prefs in order
-        # here isn't necessary
-        for keylist in [system_keylist, user_keylist]:
-            if keylist:
-                for key in keylist:
-                    prefs[key] = self._get_macos_pref(key)
-        return prefs
-
-    def _get_file_prefs(self):
-        r"""Lookup preferences for Windows in a standardized path, such as:
-        * `C:\\Users\username\AppData\Local\Autopkg\config.{plist,json}`
-        * `/home/username/.config/Autopkg/config.{plist,json}`
-        Tries to find `config.plist`, then `config.json`."""
-
-        config_dir = appdirs.user_config_dir(APP_NAME, appauthor=False)
-
-        # Try a plist config, then a json config.
-        data = self._parse_json_or_plist_file(os.path.join(config_dir, "config.plist"))
-        if data:
-            return data
-        data = self._parse_json_or_plist_file(os.path.join(config_dir, "config.json"))
-        if data:
-            return data
-
-        return {}
-
-    def _set_macos_pref(self, key, value):
-        """Sets a preference for domain"""
-        try:
-            CFPreferencesSetAppValue(key, value, BUNDLE_ID)
-            if not CFPreferencesAppSynchronize(BUNDLE_ID):
-                raise PreferenceError(f"Could not synchronize preference {key}")
-        except Exception as err:
-            raise PreferenceError(f"Could not set {key} preference: {err}") from err
-
-    def read_file(self, file_path):
-        """Read in a file and add the key/value pairs into preferences."""
-        # Determine type or file: plist or json
-        data = self._parse_json_or_plist_file(file_path)
-        for k in data:
-            self.prefs[k] = data[k]
-
-    def _write_json_file(self):
-        """Write out the prefs into JSON."""
-        try:
-            assert self.file_path is not None
-            with open(self.file_path, "w") as f:
-                json.dump(
-                    self.prefs,
-                    f,
-                    skipkeys=True,
-                    ensure_ascii=True,
-                    indent=2,
-                    sort_keys=True,
-                )
-        except Exception as e:
-            log_err(f"Unable to write out JSON: {e}")
-
-    def _write_plist_file(self):
-        """Write out the prefs into a Plist."""
-        try:
-            assert self.file_path is not None
-            with open(self.file_path, "wb") as f:
-                plistlib.dump(self.prefs, f)
-        except Exception as e:
-            log_err(f"Unable to write out plist: {e}")
-
-    def write_file(self):
-        """Write preferences back out to file."""
-        if not self.file_path:
-            # Nothing to do if we weren't given a file
-            return
-        if self.type == "json":
-            self._write_json_file()
-        elif self.type == "plist":
-            self._write_plist_file()
-
-    def get_pref(self, key):
-        """Retrieve a preference value."""
-        return deepcopy(self.prefs.get(key))
-
-    def get_all_prefs(self):
-        """Retrieve a dict of all preferences."""
-        return self.prefs
-
-    def set_pref(self, key, value):
-        """Set a preference value."""
-        self.prefs[key] = value
-        # On macOS, write it back to preferences domain if we didn't use a file
-        if is_mac() and self.type is None:
-            self._set_macos_pref(key, value)
-        elif self.file_path is not None:
-            self.write_file()
-        else:
-            log_err(f"WARNING: Preference change {key}=''{value}'' was not saved.")
-
 
 # Set the global preferences object
 globalPreferences = Preferences()
