+++ Code/tests/test_autopkglib.py
@@ -4,6 +4,7 @@ import imp
 import json
 import os
 import plistlib
+import sys
 import unittest
 from textwrap import dedent
 from unittest.mock import mock_open, patch
@@ -139,48 +140,7 @@ class TestAutoPkg(unittest.TestCase):
     def tearDown(self):
         pass
 
-    @patch("autopkglib.sys")
-    def test_is_mac_returns_true_on_mac(self, mock_sys):
-        """On macOS, is_mac() should return True."""
-        mock_sys.platform = "Darwin-somethingsomething"
-        result = autopkglib.is_mac()
-        self.assertEqual(result, True)
-
-    @patch("autopkglib.sys")
-    def test_is_mac_returns_false_on_not_mac(self, mock_sys):
-        """On not-macOS, is_mac() should return False."""
-        mock_sys.platform = "Win32-somethingsomething"
-        result = autopkglib.is_mac()
-        self.assertEqual(result, False)
-
-    @patch("autopkglib.sys")
-    def test_is_windows_returns_true_on_windows(self, mock_sys):
-        """On Windows, is_windows() should return True."""
-        mock_sys.platform = "Win32-somethingsomething"
-        result = autopkglib.is_windows()
-        self.assertEqual(result, True)
-
-    @patch("autopkglib.sys")
-    def test_is_windows_returns_false_on_not_windows(self, mock_sys):
-        """On not-Windows, is_windows() should return False."""
-        mock_sys.platform = "Darwin-somethingsomething"
-        result = autopkglib.is_windows()
-        self.assertEqual(result, False)
-
-    @patch("autopkglib.sys")
-    def test_is_linux_returns_true_on_linux(self, mock_sys):
-        """On Linux, is_linux() should return True."""
-        mock_sys.platform = "Linux-somethingsomething"
-        result = autopkglib.is_linux()
-        self.assertEqual(result, True)
-
-    @patch("autopkglib.sys")
-    def test_is_linux_returns_false_on_not_linux(self, mock_sys):
-        """On not-Linux, is_linux() should return False."""
-        mock_sys.platform = "Win32-somethingsomething"
-        result = autopkglib.is_linux()
-        self.assertEqual(result, False)
-
+    @unittest.skipUnless("win32" in sys.platform.lower(), "requires Windows")
     @patch("autopkglib.sys")
     @patch("autopkglib.is_executable")
     @patch("autopkglib.os.get_exec_path")
