+++ Code/tests/test_preferences.py
@@ -8,7 +8,7 @@ from tempfile import TemporaryDirectory
 from unittest import mock
 from unittest.mock import patch
 
-from autopkglib import Preferences
+from autopkglib.prefs import Preferences
 
 TEST_JSON_PREFS = b"""{"CACHE_DIR": "/path/to/cache"}"""
 TEST_PLIST_PREFS = b"""<?xml version="1.0" encoding="UTF-8"?>
@@ -55,13 +55,13 @@ class TestPreferences(unittest.TestCase):
         # Return an empty list of preference keys by default. Makes a new list on
         # every call to ensure no false sharing.
         self.mock_copykeylist.side_effect = lambda *_, **_kw: list()
-        self.mock_copyappvalue = patch("autopkglib.CFPreferencesCopyAppValue").start()
-        self.mock_setappvalue = patch("autopkglib.CFPreferencesSetAppValue").start()
+        self.mock_copyappvalue = patch("autopkglib.prefs.CFPreferencesCopyAppValue").start()
+        self.mock_setappvalue = patch("autopkglib.prefs.CFPreferencesSetAppValue").start()
         self.mock_appsynchronize = patch(
-            "autopkglib.CFPreferencesAppSynchronize"
+            "autopkglib.prefs.CFPreferencesAppSynchronize"
         ).start()
 
-        self.mock_appdirs = patch("autopkglib.appdirs").start()
+        self.mock_appdirs = patch("autopkglib.prefs.appdirs").start()
         # Ensure we don't accidentally load real config and muck up tests.
         self.mock_appdirs.user_config_dir.return_value = self._workdir.name
